Moved basic auth to separate file
